@@ -3,18 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
     "axios": "^0.22.0",
     "firebase": "^9.1.2",
-    "node-sass": "^6.0.1",
-    "react": "^17.0.2",
-    "react-beforeunload": "^2.5.2",
-    "react-dom": "^17.0.2",
-    "react-draggable": "^4.4.4",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
+    "node-sass": "^8.0.0",
+    "react": "^18.2.0",
+    "react-beforeunload": "^2.5.3",
+    "react-dom": "^18.2.0",
+    "react-draggable": "^4.4.5",
+    "react-router-dom": "^5.3.4",
+    "react-scripts": "^5.0.1",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
